Move command name trimming to main loop
This way, the strings passed to commands do not include the command
itself. This allows the logic to be more consistent between commands
that require specifying the command, and the default addTodo action
where the command name is optional.

I will keep the existing functionality of adding a todo by default if no
other command is specified, unless this turns out to be a bad idea as
the project evolves.